Begin developing webpage
Begin developing a simplified version to run in a webpage. This version will not implement the genetic algorithm, but it will use "brains" discovered by the python script version to run the food chasing simulation. This is my first attempt at using GreenSock. I may switch libraries depending on how this turns out.