Begin Developing a Webpage
Begin developing a webpage which will use neural networks discovered with the python script. This page will not implement a genetic algorithm; it will only demonstrate the food chasing simulation. This is my first time using the GSAP, so I may find later that another library suits this project better, but for now I am going to play around with this as the idea.